+++ libs/base/src/text/fmt/num_to_str/tests.rs
@@ -1,4 +1,4 @@
-// devela::text::fmt::num_to_str::tests
+// devela_base::text::fmt::num_to_str::tests
 
 use super::NumToStr;