+++ libs/base/src/text/mod.rs
@@ -5,11 +5,17 @@
 // safety
 #![cfg_attr(all(feature = "base_safe", feature = "safe_text"), forbid(unsafe_code))]
 
+#[allow(hidden_glob_reexports, reason = "re-exported char")]
+mod char;
+
 pub mod errors;
 pub mod fmt;
 pub mod str;
 
-crate::structural_mods! { // _pub_mods
+crate::structural_mods! { // mods, _pub_mods
+    _mods {
+        pub use super::char::_all::*;
+    }
     _pub_mods {
         #[doc(inline)]
         pub use super::errors::*;