+++ src/text/str/reexports.rs
@@ -3,11 +3,31 @@
 //! String related re-exports.
 //
 
-use crate::impl_cdef;
+use crate::{ConstDefault, impl_cdef, paste, unwrap};
 
 crate::mod_path!(+pub _c "../../../libs/base/src/text/str/reexports.rs");
 crate::mod_path!(alloc +pub _a "../../../libs/base_alloc/src/text/str/reexports.rs");
 crate::mod_path!(std +pub _s "../../../libs/base_std/src/text/str/reexports.rs");
+
+pub use devela_base::text::str::{Str, StringU8, StringU16, StringU32, StringUsize};
+
+macro_rules! impl_cdef_for_string_u {
+    () => { impl_cdef_for_string_u![u8, u16, u32, usize]; };
+    (// $t: the length type. E.g.: u8.
+    $($t:ty),+ $(,)?) => {
+        $(
+            paste! { impl_cdef_for_string_u![@[<String $t:camel>], $t]; }
+        )+
+    };
+    (// $name: the name of the type. E.g.: StringU8.
+    @$name:ty, $t:ty) => { paste! {
+        impl<const CAP: usize> ConstDefault for $name<CAP> {
+            #[doc = "Returns an empty string.\n\n#Panics\n\nPanics if `CAP > `[`" $t "::MAX`]."]
+            const DEFAULT: Self = unwrap![ok Self::new()];
+        }
+    }};
+}
+impl_cdef_for_string_u!();
 
 /* from other modules */